Refactor: Split usePlanSubscriptions hook
Split the large `usePlanSubscriptions` hook into smaller, more focused hooks to improve maintainability.  This will make the code easier to understand, test, and maintain in the long run.
